Get rid of traits module
