+++ sdk/cosmos/src/lib.rs
@@ -108,12 +108,13 @@ pub mod responses;
 mod consistency_level;
 mod errors;
 mod headers;
+mod max_item_count;
 mod partition_keys;
 mod resource_quota;
 mod to_json_vector;
-mod traits;
 
 pub use consistency_level::ConsistencyLevel;
+pub use max_item_count::MaxItemCount;
 pub use partition_keys::PartitionKeys;
 pub use resource_quota::ResourceQuota;
