+++ sdk/cosmos/src/resources/permission/mod.rs
@@ -8,3 +8,22 @@ mod permission_token;
 pub use authorization_token::AuthorizationToken;
 pub use permission::{Permission, PermissionMode};
 pub use permission_token::PermissionToken;
+
+use crate::headers;
+use azure_core::AddAsHeader;
+use http::request::Builder;
+
+#[derive(Debug, Clone, Copy)]
+pub struct ExpirySeconds(u64);
+
+impl ExpirySeconds {
+    pub fn new(secs: u64) -> Self {
+        Self(secs)
+    }
+}
+
+impl AddAsHeader for ExpirySeconds {
+    fn add_as_header(&self, builder: Builder) -> Builder {
+        builder.header(headers::HEADER_DOCUMENTDB_EXPIRY_SECONDS, self.0)
+    }
+}
